refactor(licensing): unify license refresh to one pre-register path (TIDE review)
- Delete WcpLicenseInitializer + the post-register RequestInitializer refresh.
  There is now ONE per-request refresh: `await loadWcpLicense()` at the top of
  registerApiRequestStack. WcpLicenseProvider.get() reads that process cache, so
  WcpContext.canUse* and FeatureFlags both see the live license (fresh at register
  and resolver time).
- WcpLicenseProvider: get() → getCachedWcpLicense(); an initial license (tests)
  seeds the cache via seedWcpLicenseCache(). Dropped refresh() (no callers left).
- Revert #5559's FeatureFlagsWithLicenseDecorator to read WcpLicenseProvider (the
  provider now returns the pre-register cache), so that file is untouched again.
- Extract resolveWcpProjectEnvironment() into its own file.
