+++ packages/api-core/src/features/wcp/loadWcpLicense.ts
@@ -1,44 +1,9 @@
-import { getWcpProjectLicense, getWcpProjectEnvironment, NullLicense } from "@webiny/wcp";
-import type {
-    DecryptedWcpProjectLicense,
-    ILicense,
-    WcpProjectEnvironment
-} from "@webiny/wcp/types.js";
+import { getWcpProjectLicense, NullLicense } from "@webiny/wcp";
+import type { DecryptedWcpProjectLicense, ILicense } from "@webiny/wcp/types.js";
 import { License } from "@webiny/wcp";
 import type { CachedWcpProjectLicense } from "~/features/wcp/WcpContext/types.js";
 import { getWcpProjectLicenseCacheKey } from "~/features/wcp/WcpContext/utils.js";
-
-/**
- * Resolve the WCP project environment used to fetch the license. Prefer the CLI-computed
- * WCP_PROJECT_ENVIRONMENT blob (set by `applyWcpEnvVars` during `webiny serve` and the AWS deploy).
- * When it's absent — a CLI-less self-hosted deploy that boots the built handler directly (e.g.
- * `node start.mjs`) — derive the same org, project and api key straight from the runtime env, so a
- * fresh license can still be fetched.
- */
-function resolveWcpProjectEnvironment(): WcpProjectEnvironment | null {
-    const fromBlob = getWcpProjectEnvironment();
-    if (fromBlob) {
-        return fromBlob;
-    }
-
-    const projectId = process.env.WEBINY_PROJECT_ID || process.env.WCP_PROJECT_ID;
-    const apiKey =
-        process.env.WEBINY_PROJECT_API_KEY || process.env.WCP_PROJECT_ENVIRONMENT_API_KEY;
-    if (!projectId || !apiKey) {
-        return null;
-    }
-
-    // WEBINY_PROJECT_ID has an "orgId/projectId" shape.
-    const [orgId, project] = projectId.split("/");
-    if (!orgId || !project) {
-        return null;
-    }
-
-    // `id` is intentionally empty. The WCP backend identifies the specific project environment
-    // from the api key (sent as the request's authorization header); the license fetch only uses
-    // org/project (URL path) + apiKey (auth), never the environment id. So there's nothing to fill.
-    return { id: "", apiKey, org: { id: orgId }, project: { id: project } };
-}
+import { resolveWcpProjectEnvironment } from "~/features/wcp/resolveWcpProjectEnvironment.js";
 
 const wcpProjectEnvironment = resolveWcpProjectEnvironment();
 
@@ -59,6 +24,16 @@ let inflight: Promise<ILicense> | null = null;
  */
 export function getCachedWcpLicense(): ILicense {
     return cachedLicense.license;
+}
+
+/**
+ * Seed the cache with an already-resolved license (tests pass a pre-built one). Marks it fresh so a
+ * subsequent `loadWcpLicense()` returns it without a fetch. This is how `WcpLicenseProviderImpl`
+ * installs an initial license.
+ */
+export function seedWcpLicenseCache(license: ILicense): void {
+    cachedLicense.license = license;
+    cachedLicense.cacheKey = getWcpProjectLicenseCacheKey();
 }
 
 export async function loadWcpLicense(
