+++ packages/api-core/src/features/wcp/resolveWcpProjectEnvironment.ts
@@ -1,0 +1,34 @@
+import { getWcpProjectEnvironment } from "@webiny/wcp";
+import type { WcpProjectEnvironment } from "@webiny/wcp/types.js";
+
+/**
+ * Resolve the WCP project environment used to fetch the license. Prefer the CLI-computed
+ * WCP_PROJECT_ENVIRONMENT blob (set by `applyWcpEnvVars` during `webiny serve` and the AWS deploy).
+ * When it's absent — a CLI-less self-hosted deploy that boots the built handler directly (e.g.
+ * `node start.mjs`) — derive the same org, project and api key straight from the runtime env, so a
+ * fresh license can still be fetched.
+ */
+export function resolveWcpProjectEnvironment(): WcpProjectEnvironment | null {
+    const fromBlob = getWcpProjectEnvironment();
+    if (fromBlob) {
+        return fromBlob;
+    }
+
+    const projectId = process.env.WEBINY_PROJECT_ID || process.env.WCP_PROJECT_ID;
+    const apiKey =
+        process.env.WEBINY_PROJECT_API_KEY || process.env.WCP_PROJECT_ENVIRONMENT_API_KEY;
+    if (!projectId || !apiKey) {
+        return null;
+    }
+
+    // WEBINY_PROJECT_ID has an "orgId/projectId" shape.
+    const [orgId, project] = projectId.split("/");
+    if (!orgId || !project) {
+        return null;
+    }
+
+    // `id` is intentionally empty. The WCP backend identifies the specific project environment
+    // from the api key (sent as the request's authorization header); the license fetch only uses
+    // org/project (URL path) + apiKey (auth), never the environment id. So there's nothing to fill.
+    return { id: "", apiKey, org: { id: orgId }, project: { id: project } };
+}
