+++ packages/project-server/src/serve/runApiServer.ts
@@ -63,8 +63,8 @@ export async function runApiServer(
     // Hybrid runtime env: the api allowlist (WEBINY_/WCP_PROJECT_ENVIRONMENT/OKTA_/AUTH0_ + DEBUG) —
     // the same app vars the AWS Lambda forwards, no arbitrary leakage — PLUS the system vars a spawned
     // process needs to run (PATH / HOME / NODE_EXTRA_CA_CERTS for portless TLS / temp / locale).
-    // `WCP_PROJECT_LICENSE` is absent (not allowlisted), so the handler's WcpLicenseInitializer
-    // fetches a fresh license rather than reading the build-time plaintext value.
+    // `WCP_PROJECT_LICENSE` is absent (not allowlisted), so the handler's per-request license refresh
+    // (`loadWcpLicense`) fetches a fresh license rather than reading the build-time plaintext value.
     const runtimeEnv = pickServerRuntimeEnvVariables();
 
     // The database is driver-agnostic here: the DB infra extension in webiny.config (<Infra.Sqlite> /
